Move nondet{cpp,h} to `util`
Make nondet.{h,cpp} generally available for CBMC.
